internal/giofs: add
